refactor(cold-window): split tui.rs into a module tree
The 1960-line tui.rs mixed shell state, responsive layout, frame
rendering, key routing, the event loop, and tests in one file,
flagged as too large in PR #230 review.

Split it into a tui/ module by concern: mod.rs keeps the shell
state (UiState, SelectionState) and re-exports the public surface;
layout.rs owns tier selection and pane placement; render.rs owns
draw; input.rs owns handle_key and its helpers; runner.rs owns the
terminal lifecycle and event loop; tests.rs holds the suite.

Public paths are unchanged: mod.rs re-exports draw, handle_key,
run, KeyOutcome, QuotaFn, TuiOptions, and UiState, so the server
crate's tui_flow consumer and every cold_window::tui import compile
untouched. Cross-module helpers (plan_for, GUARD_MIN_* consts) were
promoted to pub(crate); columnar stays private to layout.rs.

Pure refactor, no behavior change: the existing 32 tui tests plus
the server cold-window parity and first-paint tests all pass.
